api-app/requirements.txt
annoy==1.16.2
astroid==2.2.5
autopep8==1.4.3
Django==2.2.5
django-cors-headers==3.1.1
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
isort==4.3.16
joblib==0.14.0
lazy-object-proxy==1.3.1
lxml==4.3.3
mccabe==0.6.1
mysqlclient==1.4.4
numpy==1.17.1
opencv-python==4.1.0.25
pandas==0.25.1
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
PyQt5==5.12.2
PyQt5-sip==4.19.17
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
schedule==0.6.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
tqdm==4.35.0
typed-ast==1.3.1
uWSGI==2.0.18
wrapt==1.11.1
